requirements.txt
future==0.15.2
jira==1.0.3
oauthlib==1.0.3
python-telegram-bot==3.4
requests==2.9.1
requests-oauthlib==0.6.1
requests-toolbelt==0.6.0
six==1.10.0
tlslite==0.4.9
